// src/utils/types.ts
export type ApiResponse<T> = {
    success: boolean;
    message?: string;
    data?: T;
};
// ...
export type PastMedicalHistoryDto = {
    id?: number;
    orgId?: number;
    externalId?: string | null;

    patientId: number;
    encounterId: number;

    description: string; // e.g., "Patient has history of hypertension, diagnosed in 2018."
    esigned?: boolean;
    signedAt?: string;
    signedBy?: string;

    audit?: {
        createdDate?: string;
        lastModifiedDate?: string;
    };
};